Tagged-PDF: Heuristics in poppler-glib for data/layout table identification
Add functions in poppler-glib which, used on a PopplerStructureElement of
type POPPLER_STRUCTURE_TABLE, determine whether a table is used for layout
or contains actual data. The heuristic is quite simple so far:

- poppler_structure_element_is_data_table(): checks for table headings
  and that THead/TBody elements are present.
- poppler_structure_element_is_layout_table(): negates the result of
  the previous function.

The idea is that in the later case a different heuristic could be
implemented later on, and at some point tables could be "data tables",
"layout tables", or the heuristics can not tell for sure.