Tagged-PDF: Implement parsing of StructTreeRoot
Implement parsing of the StructTreeRoot entry of the Catalog. Also, the
Catalog::getStructTreeRoot() and PDFDoc::getStructTreeRoot() methods are
modified to return an instance of StructTreeRoot instead of an Object.

All elements from the StructTreeRoot are parsed except for:

- IDTree: it is a lookup tree to locate items by their ID, which would
  be barely useful because the whole structure tree is to be kept in
  memory, which should be fast enough to traverse.
- ParentTreeNextKey: This is needed only when the ParentTree object is
  to be modified. For the moment the implementation deals only with
  reading, so this has been deliberately left out.

Also, pdfinfo is used to print tagging info from Catalog::getMarkInfo()
instead opf assuming that the presence of the StrucTreeRoot implies that
the file is tagged.